feat(prompting-client): bundle the scripted and echo clients in the snap package
